server/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
